requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
streamlit==1.40.0
langgraph==0.2.45
openai==1.54.0
python-multipart==0.0.12
requests==2.32.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
